Implemented contexts, removal, padding, transients; improved config.
Also added the activate function; need to add setter to
defcustom for the smart equals key as that key is now in both
the minor mode map and the transient map.

Still want to add tab key to select operator by completion, but
the stub is tere.

Need to watch spacing and add completion of partial match

Want to refactor essmeq--process to extract the matching
and accept plus coordinates, as this can be reused by several
pieces here, making --process smoother at the same time.

Still several more TODOs, plus more tests and documentation.