Provisional (and possibly working) version of essmeq--partial-match
This will be the basis of a two step operation: first search for
an exact match and only if that fails, search for a partial match
to complete.  The reason for this is that partial matching has
ambiguities that get in the way of the priority searching, e.g.,
seeing << just before point should clearly complete to <<-, but
the higher priority match is <-, which would give < <-, which
is clearly wrong. The partial match respects priority among
those that are equally far back starting point. (Test this.)
This will support the refactoring of essmeq--process to separate
the matching from the replacement as discussed earlier.