Finished search, completion, removal, and selection plus cleanup.
Need to test everything and fix several little issues and ATTNs,
along with documentation fixes.

Consider an ess-smart-equals-overrriding-context which can
be set and is cleared in the on-exit function to set-transient-map.
This lets us temporarily set the context to say 'all for cycling
through all operators.  Is this worthwhile?

Also, note two known issues in context matching with ESS and their
fixes described with ess--context.  Is the fix to funcall context
good enough? Seems to work.  Adding scanning for % is easy
with temporary insertion :before

essmeq--inside-brackets-p:
(essmeq--with-temporary-insertion "%" :before
  (ess-inside-brackets-p))

essmeq--inside-call-p*: when (ess-inside-string-p) returns ?%
(essmeq--with-temporary-insertion "%" :before
  (essmeq-inside-call-p))