Fixed requires and circularity and added some tests
The use of setters in the defcustoms for the configurable
parameters was generating a bit of circularity in that
updates to several of the maps were best done with a
function that referred to several of the variables making
a linear ordering impossible.

So, I put all the updates in one function and put it in
the minor mode enable code along with an interactive
function ess-smart-equals-update-keymaps which can
always be used if needed. Documentation was updated
to address this. (One of the original setters, for the
mode map, had another small bug anyway.) I considered
caching -key, -extra-ops, and -cancel-keys to avoid
mucking with the keymaps unnecessarily, but that seemed
excessive. If there are issues as is, we can consider
that. The caching code looked like the following:

(defvar essmeq--key-cache
  (vector ess-smart-equals-key
          ess-smart-equals-extra-ops
          ess-smart-equals-cancel-keys))

(defun emacs-smart-equals-update-keymaps ()
  "Force keymap update in `ess-smart-equals-mode' to catch config changes.
This should not usually need to be done explicitly by the user."
  (unless (and (equal ess-smart-equals-key (aref essmeq--key-cache 0))
               (equal ess-smart-equals-cancel-keys (aref essmeq--key-cache 2)))
    (setf essmeq--transient-map (essmeq--make-transient-map)
          (aref essmeq--key-cache 0) ess-smart-equals-key
          (aref essmeq--key-cache 2) ess-smart-equals-cancel-keys))
  (unless (and (equal ess-smart-equals-key (aref essmeq--key-cache 0))
               (equal ess-smart-equals-extra-ops (aref essmeq--key-cache 1)))
    (setf ess-smart-equals-mode-map (essmeq--make-mode-map)
          (aref essmeq--key-cache 0) ess-smart-equals-key
          (aref essmeq--key-cache 1) ess-smart-equals-extra-ops)))