New op order resolves c/cycle ambiguity; paren improved; fixed tests
The completion/cycling ambiguity is not gone, but with
the default operator ordering, it is now papered over.
Specifically, things will be fine if each prefix string
in the list directly precedes a string it is a prefix of.
For example, "<" "<="; then on seeing < and hitting equals,
the result is cycling (since "<" is in the op list) and
the cycling gives the natural result "<=".  Any other order
will break this, and it is worth considering what can be
done. One way: distinguish cycling from completion based
on whether (a) last-command equals this command and
(b) whether starting point is after a space character.
(Condition b will only work when there is right padding.)
If cycling, do match then complete then fallback-complete;
otherwise do  complete then match then fallback-complete.
I think this will fix the ambiguity on any ordering, which is
good, but it is a bit clunky. I will try this.

Changed _ to a _ _ in the () skeleton spec because wrapping
the region would put point after the pair otherwise instead
of on the magic space.  Also coded an experimental expand
key on magic space that wraps the next expression; unclear
whether to use a token or an expression here, whether to
eliminate intervening spaces (or squash them or leave them be),
and whether to limit expansion to one line. Should also
consider whether space is the right key to use here; might get
intrusive. Do I need an option to disable this feature? It
might be too much, and should be a lower priority.

Need to handle padding properly; other than cycling/completion
ambiguity that is the big remaining issue outstanding.