Small fixes for package linter
Note that the package linter rejects
requiring ess-r-mode, which is included in ess.
I don't want to do require ess as that is not minimal,
but in the end it would probably do no harm.

Second, the package linter rejects the secondary prefix
essmeq- which is used only for internal private functions.
I'd like to keep this; hopefully, MELPA will not give a problem.